style: Added responsiveness and Enhanced the ui of community creation page
