+++ apps/web/src/app/(verified)/create-community/page.tsx
@@ -1,8 +1,10 @@
-import CreateCommunity from "@/components/chatRoom/CreateCommunity";
+import CreateCommunity from "@/components/createCommunity/CreateCommunity";
+import PicGrid from "@/components/createCommunity/PicGrid";
 
 const page = () => {
   return (
-    <div>
+    <div className="flex flex-row items-start justify-between md:py-8 py-4 mx-6 md:mx-8 md:mr-10 gap-8">
+      <PicGrid />
       <CreateCommunity />
     </div>
   );
